moved RHS function to seperate hpp
