+++ src/models/RHS.hpp
@@ -2,31 +2,51 @@
 #define RHS_HPP
 
 #include <vector>
-#include <algorithm>  // for std::max
-#include <cmath>      // for std::pow
-#include <boost/numeric/odeint.hpp>
+#include <algorithm> // for std::max
+#include <cmath>     // for std::pow
+#include <boost/numeric/odeint.hpp> // assuming you're using Boost odeint
 
 using namespace boost::numeric::odeint;
 
 struct RHS {
-    const std::vector<double>& runoff_series;
-    const std::vector<double>& y_p_series;
-    const double A_h;
-    const double lambda_1;
-    const double invtau;
+    const float* runoff_series; // runoff, hourly
+    const std::vector<double>& y_p_series; // inflow from parent nodes, hourly (for now)
+    const double A_h; // hillslope area in m^2
+    const double lambda_1; // exponent
+    const double invtau; // constant term
 
-    RHS(const std::vector<double>& runoff_series,
+    RHS(const float* runoff_series,
         const std::vector<double>& y_p_series,
-        const double A_h,
-        const double lambda_1,
+        const double A_h, 
+        const double lambda_1, 
         const double invtau
-    );
+    )
+    : runoff_series(runoff_series),
+      y_p_series(y_p_series),
+      A_h(A_h),
+      lambda_1(lambda_1),
+      invtau(invtau) {}
 
-    void operator()(const double q, double& dQdt, const double t) const;
+    inline void operator()(const double q, double& dQdt, const double t) const {
+        // t is in minutes, inputs are in hours 
+        size_t hour_idx = static_cast<size_t>(t / 60);  // assumes t in minutes 
+
+        // runoff_series is hourly 
+        double runoff = runoff_series[hour_idx] * (0.001 / 60); // mm/h to m/min
+
+        // y_p is hourly — 60 minutes per step
+        double y_p = y_p_series[hour_idx];
+
+        // Nonlinear routing ODE
+        double q_safe = std::max(q, 1e-8); // avoid division by zero
+        dQdt = invtau * pow(q_safe, lambda_1) * (-q_safe + (runoff * A_h / 60.0) + y_p);
+    }
 };
 
-// RK45 stepper type alias
-typedef runge_kutta_dopri5<double, double, double, double, vector_space_algebra, default_operations, never_resizer> stepper_type;
-extern stepper_type::controlled_type stepper;
+// RK45 Solver typedef
+typedef runge_kutta_dopri5<
+    double, double, double, double,
+    vector_space_algebra, default_operations, never_resizer
+> stepper_type;
 
 #endif // RHS_HPP
