+++ src/routing_test.cpp
@@ -10,6 +10,8 @@
 #include "I_O/node_info.hpp"
 #include "I_O/output_series.hpp"
 #include "I_O/inputs.hpp"
+#include "models/RHS.hpp"
+
 
 // C++ standard libraries
 using namespace std;
@@ -28,48 +30,47 @@ using namespace boost::numeric::odeint;
 //   - If the boundary conditions are valid for the entire simulation time.
 // 8. Move all parts of main to functions to improve readability and maintainability.
 
-//use classes for object oriented programming for all functions
 
 // ROUTING ODES--------------------------------------------------
-struct RHS {
-    const float* runoff_series; // runoff, hourly
-    const std::vector<double>& y_p_series; // inflow from parent nodes, hourly (for now)
-    const double A_h; // hillslope area in m^2
-    const double lambda_1; // exponent
-    const double invtau; // constant term
+// struct RHS {
+//     const float* runoff_series; // runoff, hourly
+//     const std::vector<double>& y_p_series; // inflow from parent nodes, hourly (for now)
+//     const double A_h; // hillslope area in m^2
+//     const double lambda_1; // exponent
+//     const double invtau; // constant term
 
-    RHS(const float* runoff_series,
-        const std::vector<double>& y_p_series,
-        const double A_h, 
-        const double lambda_1, 
-        const double invtau
-        )
-        : runoff_series(runoff_series),
-        y_p_series(y_p_series),
-        A_h(A_h),
-        lambda_1(lambda_1),
-        invtau(invtau){}
+//     RHS(const float* runoff_series,
+//         const std::vector<double>& y_p_series,
+//         const double A_h, 
+//         const double lambda_1, 
+//         const double invtau
+//         )
+//         : runoff_series(runoff_series),
+//         y_p_series(y_p_series),
+//         A_h(A_h),
+//         lambda_1(lambda_1),
+//         invtau(invtau){}
 
-    inline void operator()(const double q, double& dQdt, const double t) const {
-        // t is in minutes, inputs are in hours 
-        size_t hour_idx = t/60;  // assumes t in minutes (!!!!need to pass index for runoff and y_p series)
+//     inline void operator()(const double q, double& dQdt, const double t) const {
+//         // t is in minutes, inputs are in hours 
+//         size_t hour_idx = t/60;  // assumes t in minutes (!!!!need to pass index for runoff and y_p series)
 
-        // runoff_series is hourly 
-        double runoff = runoff_series[hour_idx]*(0.001/60); // convert from mm/h to m/min (assuming runoff is in mm/h)
+//         // runoff_series is hourly 
+//         double runoff = runoff_series[hour_idx]*(0.001/60); // convert from mm/h to m/min (assuming runoff is in mm/h)
 
-        // y_p is hourly — 60 minutes per step
-        double y_p = y_p_series[hour_idx];
+//         // y_p is hourly — 60 minutes per step
+//         double y_p = y_p_series[hour_idx];
 
-        // Nonlinear routing ODE
-        double q_safe = std::max(q, 1e-8); //set a safe minimum to avoid division by zero
-        dQdt = invtau * pow(q_safe, lambda_1) * (-q_safe + (runoff * A_h / 60.0) + y_p);
-    }
-};
+//         // Nonlinear routing ODE
+//         double q_safe = std::max(q, 1e-8); //set a safe minimum to avoid division by zero
+//         dQdt = invtau * pow(q_safe, lambda_1) * (-q_safe + (runoff * A_h / 60.0) + y_p);
+//     }
+// };
 
 
-// RK45 Solver and methodology
-// state_type = value_type = deriv_type = time_type = double
-typedef runge_kutta_dopri5< double , double , double , double , vector_space_algebra , default_operations , never_resizer > stepper_type;
+// // RK45 Solver and methodology
+// // state_type = value_type = deriv_type = time_type = double
+// typedef runge_kutta_dopri5< double , double , double , double , vector_space_algebra , default_operations , never_resizer > stepper_type;
 
 int main()
 {
